Fix regression due to precision of FDTDGrid.dl
np.single caused many extra decimal places that using built in float
as the dtype does not.

+++ gprMax/cmds_singleuse.py
@@ -93,7 +93,7 @@ class Discretisation(UserObjectSingle):
     def build(self, model, uip):
         G = model.G
         try:
-            G.dl = np.array(self.kwargs["p1"], dtype=np.single)
+            G.dl = np.array(self.kwargs["p1"], dtype=float)
         except KeyError:
             logger.exception(f"{self.__str__()} discretisation requires a point")
             raise
